Restructure folders todo-app
