+++ src/Middlewares/index.js
@@ -1,3 +1,2 @@
 export * from "./fake-backend";
-export * from "./router";
 export * from "./auth-header";
