+++ src/main.js
@@ -1,17 +1,17 @@
 import Vue from "vue";
 import VeeValidate from "vee-validate";
 import Vuex from "vuex";
-import { store } from "./_store";
-import { router } from "./_helpers";
+import { store } from "./Stores";
+import { router } from "./Routes";
 import App from "./App.vue";
 
-import vuetify from "./plugins/vuetify";
+import vuetify from "./Plugins/vuetify";
 
 Vue.use(VeeValidate);
 Vue.use(Vuex);
 Vue.config.productionTip = false;
 
-import { configureFakeBackend } from "./_helpers";
+import { configureFakeBackend } from "./Middlewares";
 configureFakeBackend();
 
 new Vue({
