--- src/views/AboutView.vue
@@ -1,5 +1,0 @@
-<template>
-	<div class="about">
-		<h1>This is an about page</h1>
-	</div>
-</template>
