--- src/views/HomeView.vue
@@ -1,15 +1,0 @@
-<template>
-	<hello-world />
-</template>
-
-<script>
-import HelloWorld from "../components/HelloWorld";
-
-export default {
-	name: "HomeView",
-
-	components: {
-		HelloWorld,
-	},
-};
-</script>
